+++ db/migrate/20260101000018_add_all_comment_notices_to_bible270_readers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddAllCommentNoticesToBible270Readers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :bible270_readers, :notify_on_all_comments, :boolean, null: false, default: false
+  end
+end